+++ CHANGELOG.md
@@ -27,6 +27,9 @@ Additions
 - Added TextureExtensions.PremultipliedCopy for textures
 
 ### MLEM.Ui
+Additions
+- Added Panel.IsVisible method to check if a child element is visible
+
 Improvements
 - Construct images in UiParser.ParseImage on the main thread to support usage with KNI
 - Create a premultiplied copy of UiParser images to support usage with KNI